A new config file format
The old config file scheme used confusing Ruby-specific concepts such as
the _class_ keyword and assignment to variables with the _=_ operator.

This new config file format looks more like a sensible config file.

  name 'Hannah Albright'
  hyphenate 'im-portant', 'in-tra-net'
  hyphenate 'Surfboard'
  recipient 'Big Lots'
  recipient 'Facebook'
  recipient 'Walmart'

It's still Ruby though, so you can use all the exciting features of the
language to make your configuration as complex as possible:

  surname = "Albright"
  name "Hannah #{surname}"
  for company in [:walmart, :facebook]
    recipient company.to_s.capitalize
  end