Promote the config/options/hash mess to a Real Object
Instead of passing around ad-hoc hashes of options, we now have a real
Config object.  There's more code now, but it should avoid more of the
same old hash-key-typo bugs, and the config-defaults code is now
factored into small methods of a few lines each, which is much easier to
scrutinize.